requirements.txt
streamlit>=1.51.1
pandas>=2.3.3
chromadb>=1.3.7
google-genai>=1.56.0
langchain>=1.2.0
langchain-core>=1.2.13
langchain-community>=0.4.1
langchain-chroma>=1.1.0
langchain-google-genai>=4.2.0
langchain-classic>=1.0.0
langchain-text-splitters>=1.1.0
pypdf>=6.4.2
typing-extensions>=4.15.0
python-dotenv